use u16

+++ src/cpu.rs
@@ -94,7 +94,7 @@ impl AddrMode {
         }
     }
 
-    pub fn operand_bytes(&self) -> u8 {
+    pub fn operand_bytes(&self) -> u16 {
         match self {
             AddrMode::Implied | AddrMode::Accumulator => 0,
             AddrMode::Immediate
@@ -461,13 +461,13 @@ impl Cpu {
 
     fn execute(&mut self, bus: &mut Bus, op: &Op, opcode: u8) {
         let operand_bytes = op.mode.operand_bytes();
-        let operands = bus.read_range_mut(self.pc.wrapping_add(1), operand_bytes as u16);
+        let operands = bus.read_range_mut(self.pc.wrapping_add(1), operand_bytes);
 
         if self.log.is_some() {
             self.log(bus, opcode, op, &operands);
         }
 
-        self.pc = self.pc.wrapping_add(1 + operand_bytes as u16);
+        self.pc = self.pc.wrapping_add(1 + operand_bytes);
         let extra_cycles = (op.execute)(self, bus, op.mode, &operands);
         let total_cycles = op.base_cycles + extra_cycles;
         self.cycle_count += total_cycles as usize;
